Scaffolding test fix
