+++ marklogic-data-hub/src/test/java/com/marklogic/hub/scaffolding/ScaffoldingTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.marklogic.hub.jupiterbased;
+package com.marklogic.hub.scaffolding;
 
 import com.marklogic.hub.HubTestBase;
 import com.marklogic.hub.error.ScaffoldingValidationException;
@@ -26,10 +26,9 @@ import com.marklogic.hub.scaffold.impl.ScaffoldingImpl;
 import com.marklogic.hub.util.FileUtil;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
+import org.custommonkey.xmlunit.XMLUnit;
 import org.junit.Before;
-import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestFactory;
+import org.junit.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
 import org.xml.sax.SAXException;
@@ -40,13 +39,10 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@RunWith(JUnitPlatform.class)
 public class ScaffoldingTest extends HubTestBase {
 
     static Path projectPath = Paths.get(PROJECT_PATH).toAbsolutePath();
@@ -56,7 +52,10 @@ public class ScaffoldingTest extends HubTestBase {
 
     @Before
     public void setup() throws IOException {
-        basicSetup();
+        XMLUnit.setIgnoreWhitespace(true);
+        deleteProjectDir();
+
+        installHubOnce();
         isMl9 = getMlMajorVersion() == 9;
     }
 
@@ -78,123 +77,6 @@ public class ScaffoldingTest extends HubTestBase {
         assertFalse(flowDir.toFile().exists());
     }
 
-    @TestFactory
-    public List<DynamicTest> generateFlowTests() {
-        List<DynamicTest> tests = new ArrayList<>();
-        allCombos((codeFormat, dataFormat, flowType, useEs) -> {
-            String flowName = "test-" + flowType.toString() + "-" + codeFormat.toString() + "-" + dataFormat.toString();
-            tests.add(DynamicTest.dynamicTest(flowName, () -> {
-                FileUtils.deleteDirectory(projectDir);
-                createFlow(codeFormat, dataFormat, flowType, true);
-            }));
-        });
-
-        return tests;
-    }
-
-    private void createFlow(CodeFormat codeFormat, DataFormat dataFormat, FlowType flowType, boolean useEsModel) {
-        String entityName = "my-fun-test";
-        String flowName = "test-" + flowType.toString() + "-" + codeFormat.toString() + "-" + dataFormat.toString();
-
-        ScaffoldingImpl scaffolding = new ScaffoldingImpl(projectDir.toString(), finalClient);
-
-        Path entityDir = scaffolding.getEntityDir(entityName);
-        assertFalse(entityDir.toFile().exists(), entityDir.toString() + " should not exist but does");
-
-        Path employeeDir = scaffolding.getEntityDir("employee");
-        assertFalse(employeeDir.toFile().exists());
-
-        scaffolding.createEntity(entityName);
-        scaffolding.createEntity("employee");
-        assertTrue(projectDir.exists());
-        assertTrue(entityDir.toFile().exists());
-        assertTrue(employeeDir.toFile().exists());
-        assertEquals(Paths.get(pluginDir.toString(), "entities", entityName), entityDir);
-
-        FileUtil.copy(getResourceStream("scaffolding-test/employee.entity.json"), employeeDir.resolve("employee.entity.json").toFile());
-        FileUtil.copy(getResourceStream("scaffolding-test/" + entityName + ".json"), entityDir.resolve(entityName + ".entity.json").toFile());
-
-        installUserModules(getHubConfig(), true);
-
-        scaffolding.createFlow(entityName, flowName, flowType, codeFormat, dataFormat, useEsModel);
-        Path flowDir = scaffolding.getFlowDir(entityName, flowName, flowType);
-        assertEquals(Paths.get(pluginDir.toString(), "entities", entityName, flowType.toString(), flowName), flowDir);
-        assertTrue(flowDir.toFile().exists());
-
-        Path flowDescriptor = flowDir.resolve(flowName + ".properties");
-        assertTrue(flowDescriptor.toFile().exists());
-
-        Properties properties = new Properties();
-        try {
-            FileInputStream fis = new FileInputStream(flowDescriptor.toFile());
-            properties.load(fis);
-            fis.close();
-        } catch(IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        int expectedPropertiesCount = flowType.equals(FlowType.INPUT) ? 4 : 6;
-        assertEquals(expectedPropertiesCount, properties.keySet().size());
-        assertEquals(dataFormat.toString(), properties.get("dataFormat"));
-        assertEquals(codeFormat.toString(), properties.get("codeFormat"));
-        if (flowType.equals(FlowType.HARMONIZE)) {
-            assertEquals(codeFormat.toString(), properties.get("collectorCodeFormat"));
-            assertEquals("collector." + codeFormat.toString(), properties.get("collectorModule"));
-        }
-        assertEquals(codeFormat.toString(), properties.get("mainCodeFormat"));
-        assertEquals("main." + codeFormat.toString(), properties.get("mainModule"));
-
-        Path defaultCollector = flowDir.resolve("collector." + codeFormat.toString());
-        if (flowType.equals(FlowType.INPUT)) {
-            assertFalse(defaultCollector.toFile().exists());
-        }
-        else {
-            assertTrue(defaultCollector.toFile().exists());
-        }
-
-        Path defaultContent = flowDir.resolve("content." + codeFormat.toString());
-        assertTrue(defaultContent.toFile().exists());
-
-        if (useEsModel) {
-            try {
-                assertEquals(
-                    getResource("scaffolding-test/es-" + flowType.toString() + "-content." + codeFormat.toString())
-                        .replaceAll("\\s+", " ")
-                        .replaceAll("[\r\n]", ""),
-                    FileUtils.readFileToString(defaultContent.toFile())
-                        .replaceAll("\\s+", " ")
-                        .replaceAll("[\r\n]", ""));
-            }
-            catch(IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        Path defaultHeaders = flowDir.resolve("headers." + codeFormat.toString());
-        assertTrue(defaultHeaders.toFile().exists());
-
-        Path triplesContent = flowDir.resolve("triples." + codeFormat.toString());
-        assertTrue(triplesContent.toFile().exists());
-
-        Path writer = flowDir.resolve("writer." + codeFormat.toString());
-        if (flowType.equals(FlowType.INPUT)) {
-            assertFalse(writer.toFile().exists());
-        }
-        else {
-            assertTrue(writer.toFile().exists());
-        }
-
-        Path main = flowDir.resolve("main." + codeFormat.toString());
-        assertTrue(main.toFile().exists());
-    }
-
-    private void createInputFlow(CodeFormat codeFormat, DataFormat dataFormat, boolean useEsModel) throws IOException, SAXException {
-        createFlow(codeFormat, dataFormat, FlowType.INPUT, useEsModel);
-    }
-
-    private void createHarmonizeFlow(CodeFormat codeFormat, DataFormat dataFormat, boolean useEsModel) throws IOException, SAXException {
-        createFlow(codeFormat, dataFormat, FlowType.HARMONIZE, useEsModel);
-    }
 
     @Test
     public void createXqyRestExtension() throws IOException {
@@ -359,3 +241,4 @@ public class ScaffoldingTest extends HubTestBase {
 
 
 }
+
